DIG-1169, DIG-1402: Opa uses Vault for secret storage

+++ entrypoint.sh
@@ -2,6 +2,8 @@
 
 set -Euo pipefail
 
+OPA_ROOT_TOKEN=$(cat /run/secrets/opa-root-token)
+
 if [[ -f "/app/initial_setup" ]]; then
     sed -i s/CLIENT_ID/$KEYCLOAK_CLIENT_ID/ /app/permissions_engine/idp.rego && sed -i s/CLIENT_ID/$KEYCLOAK_CLIENT_ID/ /app/permissions_engine/authz.rego
     sed -i s/OPA_SITE_ADMIN_KEY/$OPA_SITE_ADMIN_KEY/ /app/permissions_engine/idp.rego && sed -i s/OPA_SITE_ADMIN_KEY/$OPA_SITE_ADMIN_KEY/ /app/permissions_engine/authz.rego
@@ -9,14 +11,29 @@ if [[ -f "/app/initial_setup" ]]; then
     OPA_SERVICE_TOKEN=$(cat /run/secrets/opa-service-token)
     sed -i s/OPA_SERVICE_TOKEN/$OPA_SERVICE_TOKEN/ /app/permissions_engine/authz.rego
 
-    OPA_ROOT_TOKEN=$(cat /run/secrets/opa-root-token)
     sed -i s/OPA_ROOT_TOKEN/$OPA_ROOT_TOKEN/ /app/permissions_engine/authz.rego
-    echo "initializing idp"
-    python3 /app/permissions_engine/initialize_idp.py
-    rm /app/initial_setup
+
+    echo "initializing stores"
+    python3 /app/initialize_vault_store.py
+    if [[ $? -eq 0 ]]; then
+        rm /app/initial_setup
+        rm /app/bearer.txt
+        echo "setup complete"
+    else
+        echo "!!!!!! INITIALIZATION FAILED, TRY AGAIN !!!!!!"
+    fi
 fi
+
 
 while [ 0 -eq 0 ]
 do
-  sleep 60
+  echo "storing vault token"
+  python3 get_vault_store_token.py
+  if [[ $? -eq 0 ]]; then
+      echo "vault token stored"
+      sleep 300
+  else
+      echo "vault token not stored"
+      sleep 30
+  fi
 done
